Revert `src/main` changes
